+++ core/include/core/scene/vertexes.h
@@ -7,7 +7,7 @@
 struct VertexP {
 	dg::float3 position;
 
-    static uint32_t GetVDeclId();
+    static uint16_t GetVDeclId();
 };
 
 struct VertexPNC {
@@ -15,5 +15,5 @@ struct VertexPNC {
 	dg::float3 normal;
 	dg::float2 uv;
 
-    static uint32_t GetVDeclId();
+    static uint16_t GetVDeclId();
 };
